+++ src/extra/others/fragment/lv_fragment.c
@@ -11,199 +11,9 @@
 
 #if LV_USE_FRAGMENT
 
-#include "../../../extra/widgets/msgbox/lv_msgbox.h"
-
-/*********************
- *      DEFINES
- *********************/
-
-/**********************
- *      TYPEDEFS
- **********************/
-
-typedef struct manager_stack_t {
-    const lv_fragment_class_t * cls;
-    lv_fragment_t * instance;
-    bool obj_created;
-    bool destroying_obj;
-    bool dialog;
-    struct manager_stack_t * prev;
-} manager_stack_t;
-
-struct _lv_fragment_manager_t {
-    lv_obj_t * container;
-    lv_fragment_t * parent;
-    manager_stack_t * top;
-};
-
-/**********************
- *  STATIC PROTOTYPES
- **********************/
-
-static manager_stack_t * item_new(const lv_fragment_class_t * cls);
-
-static lv_fragment_t * item_create_fragment(lv_fragment_manager_t * manager, manager_stack_t * item,
-                                            void * args);
-
-static void item_create_obj(manager_stack_t * item, lv_obj_t * parent, const lv_obj_class_t * check_type);
-
-static void item_del_obj(manager_stack_t * item);
-
-static void item_del_fragment(manager_stack_t * item);
-
-static void obj_cb_delete(lv_event_t * event);
-
-static void fragment_destroy_obj(lv_fragment_t * fragment);
-
-/**********************
- *  STATIC VARIABLES
- **********************/
-
-/**********************
- *      MACROS
- **********************/
-
-
-
 /**********************
  *   GLOBAL FUNCTIONS
  **********************/
-
-lv_fragment_manager_t * lv_fragment_manager_create(lv_obj_t * container, lv_fragment_t * parent)
-{
-    LV_ASSERT(container);
-    lv_fragment_manager_t * instance = lv_mem_alloc(sizeof(lv_fragment_manager_t));
-    instance->parent = parent;
-    instance->container = container;
-    instance->top = NULL;
-    return instance;
-}
-
-void lv_fragment_manager_del(lv_fragment_manager_t * manager)
-{
-    LV_ASSERT_NULL(manager);
-    manager_stack_t * top = manager->top;
-    while(top) {
-        LV_ASSERT(top->cls);
-        item_del_obj(top);
-        item_del_fragment(top);
-        struct manager_stack_t * prev = top->prev;
-        lv_mem_free(top);
-        top = prev;
-    }
-    lv_mem_free(manager);
-}
-
-void lv_fragment_manager_push(lv_fragment_manager_t * manager, lv_fragment_t * fragment)
-{
-    LV_ASSERT(manager);
-    LV_ASSERT(fragment);
-    LV_ASSERT(fragment->priv_item == NULL);
-    LV_ASSERT(fragment->manager == NULL);
-    manager_stack_t * item = item_new(fragment->cls);
-    lv_obj_t * parent = manager->container;
-    item->instance = fragment;
-    fragment->priv_item = item;
-    fragment->manager = manager;
-    /* Destroy object of previous screen */
-    if(manager->top) {
-        item_del_obj(manager->top);
-    }
-    item_create_obj(item, parent, NULL);
-    manager_stack_t * top = manager->top;
-    item->prev = top;
-    manager->top = item;
-}
-
-void lv_fragment_manager_replace(lv_fragment_manager_t * manager, lv_fragment_t * fragment)
-{
-    LV_ASSERT(manager);
-    LV_ASSERT(fragment);
-    LV_ASSERT(fragment->priv_item == NULL);
-    LV_ASSERT(fragment->manager == NULL);
-    manager_stack_t * top = item_new(fragment->cls);
-    top->instance = fragment;
-    fragment->priv_item = top;
-    fragment->manager = manager;
-    manager_stack_t * old = manager->top;
-    if(old) {
-        item_del_obj(old);
-        item_del_fragment(old);
-        lv_mem_free(old);
-    }
-    manager->top = top;
-    item_create_obj(top, manager->container, NULL);
-}
-
-size_t lv_fragment_manager_get_size(lv_fragment_manager_t * manager)
-{
-    size_t size = 0;
-    manager_stack_t * cur;
-    for(cur = manager->top; cur; cur = cur->prev) {
-        size++;
-    }
-    return size;
-}
-
-void lv_fragment_manager_pop(lv_fragment_manager_t * manager)
-{
-    LV_ASSERT(manager);
-    manager_stack_t * top = manager->top;
-    if(!top) return;
-    manager_stack_t * prev = top->prev;
-    bool dialog = top->dialog;
-    if(!dialog && prev) {
-        item_create_fragment(manager, prev, NULL);
-    }
-    item_del_obj(top);
-    lv_mem_free(top);
-    if(!dialog && prev) {
-        item_create_obj(prev, manager->container, NULL);
-    }
-    manager->top = prev;
-}
-
-bool lv_fragment_manager_dispatch_event(lv_fragment_manager_t * manager, int which, void * data1, void * data2)
-{
-    LV_ASSERT(manager);
-    manager_stack_t * top = manager->top;
-    if(!top) return false;
-    lv_fragment_t * instance = top->instance;
-    if(!instance || !top->cls->event_cb) return false;
-    return top->cls->event_cb(instance, which, data1, data2);
-}
-
-#if LV_USE_MSGBOX
-
-void lv_fragment_manager_show(lv_fragment_manager_t * manager, lv_fragment_t * fragment)
-{
-    LV_ASSERT(manager);
-    LV_ASSERT(fragment);
-    const lv_fragment_class_t * cls = fragment->cls;
-    manager_stack_t * item = item_new(cls);
-    item_create_obj(item, NULL, &lv_msgbox_class);
-    item->dialog = true;
-    manager_stack_t * top = manager->top;
-    item->prev = top;
-    manager->top = item;
-}
-
-#endif
-
-lv_fragment_t * lv_fragment_manager_get_top(lv_fragment_manager_t * manager)
-{
-    LV_ASSERT(manager);
-    manager_stack_t * top = manager->top;
-    if(!top)return NULL;
-    return top->instance;
-}
-
-lv_fragment_t * lv_fragment_manager_get_parent(lv_fragment_manager_t * manager)
-{
-    LV_ASSERT(manager);
-    return manager->parent;
-}
-
 
 lv_fragment_t * lv_fragment_create(const lv_fragment_class_t * cls, void * args)
 {
@@ -248,148 +58,12 @@ void lv_fragment_del_obj(lv_fragment_t * fragment)
     if(cls->obj_will_delete_cb) {
         del_handled = cls->obj_will_delete_cb(fragment, fragment->obj);
     }
-    LV_ASSERT(del_handled || fragment->obj);
+    LV_ASSERT_MSG(del_handled || fragment->obj, "Either return created object, or implement obj_will_delete_cb.");
     if(fragment->obj) {
         lv_obj_del(fragment->obj);
     }
     if(cls->obj_deleted_cb) {
         cls->obj_deleted_cb(fragment, fragment->obj);
-    }
-}
-
-void lv_fragment_recreate_obj(lv_fragment_t * fragment)
-{
-    LV_ASSERT(fragment);
-    // Disable CB first
-    lv_obj_remove_event_cb(fragment->obj, obj_cb_delete);
-    lv_fragment_del_obj(fragment);
-
-    const lv_fragment_class_t * cls = fragment->cls;
-    lv_obj_t * obj = cls->create_obj_cb(fragment, fragment->manager->container);
-    fragment->obj = obj;
-    if(cls->obj_created_cb) {
-        cls->obj_created_cb(fragment, obj);
-    }
-    if(obj) {
-        lv_obj_add_event_cb(obj, obj_cb_delete, LV_EVENT_DELETE, fragment->priv_item);
-    }
-}
-
-void lv_fragment_pop(lv_fragment_t * fragment)
-{
-    LV_ASSERT(fragment);
-    lv_fragment_manager_t * manager = fragment->manager;
-    LV_ASSERT(manager);
-    LV_ASSERT(manager->top->instance == fragment);
-    lv_fragment_manager_pop(manager);
-}
-
-/**********************
- *   STATIC FUNCTIONS
- **********************/
-
-static manager_stack_t * item_new(const lv_fragment_class_t * cls)
-{
-    LV_ASSERT(cls->instance_size);
-    manager_stack_t * item = lv_mem_alloc(sizeof(manager_stack_t));
-    lv_memset_00(item, sizeof(manager_stack_t));
-    item->cls = cls;
-    return item;
-}
-
-static lv_fragment_t * item_create_fragment(lv_fragment_manager_t * manager, manager_stack_t * item,
-                                            void * args)
-{
-    if(item->instance) return item->instance;
-    const lv_fragment_class_t * cls = item->cls;
-    LV_ASSERT(cls->instance_size);
-    lv_fragment_t * instance = lv_mem_alloc(cls->instance_size);
-    lv_memset_00(instance, cls->instance_size);
-    instance->cls = cls;
-    instance->manager = manager;
-    instance->priv_item = item;
-    if(cls->constructor_cb) {
-        cls->constructor_cb(instance, args);
-    }
-    item->instance = instance;
-    return instance;
-}
-
-static void item_create_obj(manager_stack_t * item, lv_obj_t * parent, const lv_obj_class_t * check_type)
-{
-    LV_ASSERT(item->instance);
-    lv_fragment_create_obj(item->instance, parent);
-    if(check_type) {
-        LV_ASSERT(lv_obj_has_class(item->instance->obj, check_type));
-    }
-    item->obj_created = true;
-}
-
-static void item_del_obj(manager_stack_t * item)
-{
-    if(!item->obj_created) return;
-    item->destroying_obj = true;
-    lv_fragment_t * instance = item->instance;
-    lv_fragment_del_obj(instance);
-    item->obj_created = false;
-}
-
-static void item_del_fragment(manager_stack_t * item)
-{
-    lv_fragment_t * instance = item->instance;
-    lv_fragment_del(instance);
-}
-
-static void fragment_destroy_obj(lv_fragment_t * fragment)
-{
-    const lv_fragment_class_t * cls = fragment->cls;
-    lv_obj_t * obj = fragment->obj;
-    if(obj) {
-        bool delete_handled = false;
-        if(cls->obj_will_delete_cb) {
-            delete_handled = cls->obj_will_delete_cb(fragment, obj);
-        }
-        if(!delete_handled) {
-            lv_obj_del(obj);
-        }
-        else if(cls->obj_deleted_cb) {
-            cls->obj_deleted_cb(fragment, NULL);
-        }
-    }
-    else {
-        LV_ASSERT(fragment->manager);
-        LV_ASSERT(fragment->manager->container);
-        bool delete_handled = false;
-        if(cls->obj_will_delete_cb) {
-            delete_handled = cls->obj_will_delete_cb(fragment, NULL);
-        }
-        if(!delete_handled) {
-            lv_obj_clean(fragment->manager->container);
-        }
-        if(cls->obj_deleted_cb) {
-            cls->obj_deleted_cb(fragment, NULL);
-        }
-    }
-    fragment->obj = NULL;
-}
-
-static void obj_cb_delete(lv_event_t * event)
-{
-    manager_stack_t * item = lv_event_get_user_data(event);
-    lv_fragment_t * instance = item->instance;
-    lv_fragment_manager_t * manager = instance->manager;
-    const lv_fragment_class_t * cls = instance->cls;
-    if(event->target != instance->obj) return;
-    if(cls->obj_deleted_cb) {
-        cls->obj_deleted_cb(instance, event->target);
-    }
-    item->obj_created = false;
-    instance->obj = NULL;
-    if(!item->destroying_obj) {
-        manager_stack_t * prev = item->prev;
-        item_del_fragment(item);
-        lv_mem_free(item);
-        manager->top = prev;
     }
 }
